22 add filters, error_messages, fix async 11-tg_parser

+++ app/services/telegram/telegram_channels/views.py
@@ -1,5 +1,5 @@
 from django.views import View
-from django.http import JsonResponse
+from django.http import JsonResponse, Http404
 from django.shortcuts import get_object_or_404
 from .models import Channel
 from .form import ChannelForm
@@ -10,25 +10,46 @@ from app.services.telegram.telegram_channels.utils.save_data import SaveDataChan
 from django.views.decorators.csrf import csrf_exempt
 from django.utils.decorators import method_decorator
 
+
 # Create your views here.
 
 class IndexChannelView(View):
     def get(self, request, *args, **kwargs):
-        channels = Channel.objects.all().values(
+
+        qs = Channel.objects.all()
+
+        status = request.GET.get('status')
+        if status in ['active', 'error']:
+            qs = qs.filter(status=status)
+
+        username = request.GET.get('username')
+        if username:
+            qs = qs.filter(username__icontains=username)
+
+        qs = qs.order_by('username')
+
+        channels = qs.values(
             'id', 'username', 'channel_id', 'status', 'last_message_id')
         return JsonResponse(list(channels), safe=False)
 
 
 class ShowChannelView(View):
     def get(self, request, *args, **kwargs):
-        channel = get_object_or_404(Channel, id=kwargs['pk'])
-        return JsonResponse({
-            'id': channel.id,
-            'username': channel.username,
-            'channel_id': channel.channel_id,
-            'status': channel.status,
-            'last_message_id': channel.last_message_id,
-        })
+        try:
+            channel = get_object_or_404(Channel, id=kwargs['pk'])
+            return JsonResponse({
+                'id': channel.id,
+                'username': channel.username,
+                'channel_id': channel.channel_id,
+                'status': channel.status,
+                'last_message_id': channel.last_message_id,
+            })
+        except Http404 as e:
+            return JsonResponse({
+                'status': 'error',
+                'error': 'Channel not found',
+                'details': str(e)
+            }, status=404)
 
 @method_decorator(csrf_exempt, name='dispatch')
 class AddChannelView(View):
@@ -36,9 +57,8 @@ class AddChannelView(View):
     async def get(self, request, *args, **kwargs):
         form = ChannelForm()
         return JsonResponse({
-            'status': 'error',
-            'error': form.errors,
-            'data': form.data
+            'status': 'ok',
+            'form_fields': list(form.fields.keys())
         })
 
 
@@ -48,7 +68,6 @@ class AddChannelView(View):
         # Проверка в Telegram (асинхронно)
         client_wrapper = await TelegramChannelClient.create()
         client = client_wrapper.client
-
         username = data.get('username')
         exist = ExistsTelegramChannel()
         exists = await exist.check_channel_exists(client, username)
@@ -67,8 +86,31 @@ class AddChannelView(View):
 
 @method_decorator(csrf_exempt, name='dispatch')
 class DeleteChannelView(View):
-    def post(self, request, *args, **kwargs):
+
+    def get(self, request, *args, **kwargs):
         channel_id = kwargs.get('pk')
         channel = get_object_or_404(Channel, id=channel_id)
-        channel.delete()
-        return JsonResponse({'status': 'deleted'})
+        return JsonResponse({
+            'status': 'confirm',
+            'message': f'Are you sure you want to delete the channel {channel.username}?',
+            'channel_id': channel.id
+        })
+
+    def post(self, request, *args, **kwargs):
+        try:
+            confirm = request.POST.get('confirm')
+            if confirm != 'yes':
+                return JsonResponse({
+                    'status': 'cancelled',
+                    'details': 'Deleting was calcelled by user'
+                })
+            channel_id = kwargs.get('pk')
+            channel = get_object_or_404(Channel, id=channel_id)
+            channel.delete()
+            return {'status': 'ok', 'details': 'The channel was deleted'}
+        except Http404 as e:
+            return JsonResponse({
+                'status': 'error',
+                'error': 'Channel not found',
+                'details': str(e)
+            }, status=404)
